pyproject.toml
[project]
name = "printfix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiocache>=0.12.3",
    "aiohttp>=3.13.3",
    "asgi-correlation-id>=4.3.4",
    "fastapi>=0.128.4",
    "fastmcp>=3.0.0b1",
    "google-genai>=1.62.0",
    "loguru>=0.7.3",
    "pillow>=12.0.0",
    "python-ulid>=3.1.0",
    "redis>=7.1.0",
]
